Move `admin speedtest` to `support perf` (#3966)
related to drives and objects sub-commands
