+++ cmd/auto-complete.go
@@ -406,6 +406,7 @@ var completeCmds = map[string]complete.Predictor{
 	"/support/profile/start": aliasCompleter,
 	"/support/profile/stop":  aliasCompleter,
 	"/support/inspect":       aliasCompleter,
+	"/support/perf":          aliasCompleter,
 
 	"/update": nil,
 }
